src/backend/requirements.txt
anyio==4.9.0
apify_client==1.11.0
apify_shared==1.4.1
certifi==2025.6.15
colorama==0.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
more-itertools==10.7.0
python-dotenv==1.1.1
sniffio==1.3.1
